% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAPP.R
\docType{data}
\name{convmat.23}
\alias{convmat.23}
\title{Amino acid conversion matrix with 23 properties}
\format{A data frame with 20 rows (amino acids) and 23 columns (properties)}
\usage{
data(convmat.23)
}
\description{
Conversion matrix used to transform n-mer ligand sequences into 23 * n position-by-property features.
3 properties are physicochemical: molecular weight, surface area, hydrophobicity index, and isoelectric point.
20 properties describe identity of the amino acid at a position.
}
\references{
\url{to do}
}
\keyword{datasets}
