% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hla.name.R
\name{hla.name}
\alias{hla.name}
\title{Simplify two-position HLA allotype names}
\usage{
hla.name(allele)
}
\arguments{
\item{allele}{character representing the hla name}
}
\value{
new abreviated form of the hla name that is safe to use in file names
}
\description{
\code{hla.name()} takes a proper HLA name and simplifies it down to a two position abbreviation
to get rid of special characters
}
