% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAPP.heatmap.R
\name{LAPP.heatmap}
\alias{LAPP.heatmap}
\title{Plot or save heatmap for a matrix}
\usage{
LAPP.heatmap(data.mat, min, max, do.cluster = FALSE, file.dir = NA)
}
\arguments{
\item{min}{minimum value to be used on the heatmap}

\item{max}{maximum value to be used on the heatmap}

\item{do.cluster}{boolean to determine whether the heatmaps should be clustered}

\item{file.dir}{file for the heatmap to be plotted as a pdf onto. If left empty, the heatmap
is plotted to the console instead.}

\item{matrix}{data to be plotted}
}
\value{
list of pairwise distance matrices for each feature, as well as a cumulative one
}
\description{
\code{LAPP.heatmap()} creates a heatmap using pheatmap() for either plotting or saving
}
