% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAPP.PCA.R
\name{LAPP.PCA}
\alias{LAPP.PCA}
\title{Perform PCA on HLA features}
\usage{
LAPP.PCA(features)
}
\arguments{
\item{features}{data.frame object returned by LAPP.arr.to.df().}
}
\value{
a modified PCA object from FactoMineR using the data from \code{features}
}
\description{
\code{LAPP.pca()} takes in a data.frame from LAPP.arr.to.df() and performs PCA.
The results are a modified form of results from PCA() from the FactoMineR package.
Additional information tagged onto the PCA() helps transform new data
}
