% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAPP.plot.PCA.R
\name{LAPP.plot.PCA}
\alias{LAPP.plot.PCA}
\title{Plot LAPP PCA}
\usage{
LAPP.plot.PCA(pca, dim1 = 1, dim2 = 2, type = "density",
  colors = c("#000000", "#E69F00", "#56B4E9", "#009E73", "#F0E442", "#0072B2",
  "#D55E00", "#CC79A7"))
}
\arguments{
\item{pca}{modified PCA object returned by LAPP.PCA(). Can take up to 8 different HLA alleles.}

\item{dim1}{Index of PC to plot along the x-axis. Defaults to 1}

\item{dim2}{Index of PC to plot along the y-axis. Defaults to 2}

\item{type}{Describes the type of plot to make.
Either 'points' (default) for scatter plo,t or 'density' for a 2D KDE density graph.}

\item{type}{How the PCA should be plotted. Either 'points' for scatterplot or
'density' for an automated 2D KDE plot (default).}
}
\description{
\code{LAPP.plot.PCA()} takes in a modified PCA object returned by LAPP.PCA() and plots it.
Uses ggplot2.
}
