% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAPP.plot.PCA.var.R
\name{LAPP.plot.PCA.var}
\alias{LAPP.plot.PCA.var}
\title{Plot variances for LAPP PCA}
\usage{
LAPP.plot.PCA.var(pca, top.per = 100, n = NA)
}
\arguments{
\item{pca}{modified PCA object returned by LAPP.PCA().}

\item{n}{number of eigenvalues to use. (optional)}

\item{top.perc}{top percentage of variance explained to display on graph. Default is 90\%.}
}
\description{
\code{LAPP.plot.PCA.var()} takes in a modified PCA object returned by LAPP.PCA() and plots the
 percentage of variance explained by each principal component.
 Uses ggplot2.
}
