% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAPP.plot.tSNE.R
\name{LAPP.plot.tSNE}
\alias{LAPP.plot.tSNE}
\title{Plot LAPP tSNE}
\usage{
LAPP.plot.tSNE(tsne, type = "density", colors = c("#000000", "#E69F00",
  "#56B4E9", "#009E73", "#F0E442", "#0072B2", "#D55E00", "#CC79A7"),
  clust.dict = NA)
}
\arguments{
\item{tsne}{modified PCA object returned by LAPP.tsne(). Can take up to 8 different HLA alleles.}

\item{type}{How the PCA should be plotted. Either 'points' for scatterplot or
'density' for an automated 2D KDE plot (default).}

\item{colors}{list of colors for coloring different HLA in the plot. Defaults to a colorblind palette with 8 colors.}

\item{clust.dict}{optional argument for clustering HLAs together for coloring.}
}
\description{
\code{LAPP.plot.tSNE()} takes in a list object returned by LAPP.tSNE() and plots it.
Uses ggplot2.
}
