% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAPP.R
\docType{package}
\name{LAPP}
\alias{LAPP}
\alias{LAPP-package}
\title{LAPP}
\description{
Ligand Analysis with Physicochemical Properties (LAPP) uses position by property features of HLA ligand sequences for their statistical analysis.
The package includes functions to translate HLA ligand sequences into pseudo-continuous features.
LAPP also provides statistical tests to assess how ligand sequences from different HLA differ from one another, as well as from random non-ligand peptide sequeces.
The package also includes wrapper code to create PCA scatterplots and heatmaps.
}
