% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAPP.tSNE.R
\name{LAPP.tSNE}
\alias{LAPP.tSNE}
\title{Perform tSNE on HLA features}
\usage{
LAPP.tSNE(features, iters = 1500, perplexity = 50)
}
\arguments{
\item{features}{data.frame object returned by make.features.list().}

\item{iterations}{Number of iterations for tSNE. Larger datasets mean longer runtime. Default is 1000.}
}
\description{
\code{LAPP.tSNE()} takes in a data.frame from make.features.list() and performs tSNE.
The results are a modified form of results from Rtsne() from the Rtsne package.
}
