% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAPP.wass.1.R
\name{LAPP.wass.1}
\alias{LAPP.wass.1}
\title{Calculate the cumulative first Wasserstein metric over features of peptidomes}
\usage{
LAPP.wass.1(features.df, plot.dir = NA)
}
\arguments{
\item{features.df}{features data.frame from LAPP.arr.to.df()}

\item{plot.dir}{file directory in which to save all plots}
}
\value{
list of pairwise distance matrices for each feature, as well as a cumulative one
}
\description{
\code{LAPP.wass.1()} calculates pairwise first Wasserstein metric for each of the features
of a peptidome dataset on all HLA, and also creates a final cumulative Wasserstein distance matrix using the
Euclidean distance across all metrics calculated. The function will output all distance matrices calculated.
If requested, the function will also save pdf files of heatmaps of all matrices to a file directory.
}
