% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.features.list.R
\name{make.features.list}
\alias{make.features.list}
\title{Create list of feature arrays using a list of list of ligands and a conversion matrix using make.features().}
\usage{
make.features.list(ligands.list, AA.tab, n.pos)
}
\arguments{
\item{ligands.list}{List of list of HLA allele ligand sequences created by read.lig.files().}

\item{AA.tab}{Feature conversion matrix.}

\item{n.pos}{Length of peptides (number of amino acids) in the list of ligands \code{ligands}
to convert into features. Defaults to nonamers (9).}
}
\value{
a list of a three-dimensional arrays with peptide sequence, amino acid position,
and amino acid property as the dimensions
}
\description{
\code{make.features.list()} takes in a list of ligands from read.ligand.files() and
returns a list of feature arrays using make.features().
}
