% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.features.R
\name{make.features}
\alias{make.features}
\title{Create a converted three-dimensional feature array using a list of ligands and a conversion matrix}
\usage{
make.features(ligands, AA.tab, n.pos = 9)
}
\arguments{
\item{ligands}{list of HLA peptide sequences from one HLA allele.}

\item{AA.tab}{feature conversion matrix.}

\item{n.pos}{length of peptides (number of amino acids) in the list of ligands \code{ligands}
to convert into features. Defaults to nonamers (9).}
}
\value{
a three-dimensional array with peptide sequence, amino acid position, and amino acid property
as the dimensions
}
\description{
\code{make.features()} takes in a list of ligands from one HLA and returns
a three dimensional converted feature array of the HLA's ligands using a conversion matrix.
The dimensions in order are sequence, position, and property.
}
