% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAPP.R
\docType{data}
\name{rand.9mers}
\alias{rand.9mers}
\title{Randomly selected nonamer sequences from Uniprot}
\format{A character with 100,000 entries}
\usage{
data(rand.9mers)
}
\description{
100,000 nonamer (9mer) amino acid sequences sampled from the entire Uniprot database (October 2016).
These sequences serve as useful null controls for statistical analyses in LAPP.
}
\keyword{datasets}
