% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.ligand.files.R
\name{read.ligand.files}
\alias{read.ligand.files}
\title{Read in all .txt files in a folder as ligands}
\usage{
read.ligand.files(ligand.dir, first.line.name = TRUE)
}
\arguments{
\item{ligand.dir}{file directory with one or more files of txt files of peptide sequences.
Each file is named after the HLA allele, and one peptide sequence is present per line.}

\item{first.line.name}{whether to use the first line of the file as the name of the collection of peptides
in the file (default). If set to FALSE, uses the name of the file (without the ".txt" extension) instead.}
}
\value{
list of lists of ligands in ligand.dir, where each sublist is named by each file
}
\description{
\code{read.ligand.files()} takes in a file directory and returns a list of lists of each HLA's ligands
}
