% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renormalize.R
\name{renormalize}
\alias{renormalize}
\title{Normalize feature data and drop features according to how another dataset was normalized}
\usage{
renormalize(data, info)
}
\arguments{
\item{data}{data.frame with at least one numeric column}

\item{info}{normalizing info given from normalize()}
}
\value{
new list with data.frame with normalized numeric columns and dropped constant numeric columns
and other information used to scale the data
}
\description{
\code{renormalize()} takes a data.frame and normalizes all numeric columns, as well as
dropping some, according to how a previous dataset was normalized through normalize().
}
