% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wass.1.R
\name{wass.1}
\alias{wass.1}
\title{Calculate the first Wasserstein metric (Earth Mover's Distance) of two samples}
\usage{
wass.1(x, y)
}
\arguments{
\item{x}{the first sample}

\item{y}{the second sample}
}
\value{
the first Wasserstein distance
}
\description{
\code{wass.1()} takes in two numeric vectors and returns a single numeric distance metric. The
metric is equivalent to the sum of the absolute difference in area between the empirical distribution functions
of the two samples. This total area is found by calculating left Riemann sums.
}
